requirements.txt
fastapi
uvicorn[standard]
boto3
python-dotenv
pydantic
mangum
sentence-transformers
faiss-cpu
numpy
